Update functions.py
Adding function sundata which returns data about if it's day or night in the location set in the config file, and how many percent of the day/night that has passed. Also modified the write_weather function to use that data to modify the humidity reading to write a more correct weather to database. This is done since the humidity normally is much higher in the night, without that meaning it's raining or cloudy just because of that. It does need tuning to be accurate though..